requirements.txt
flask==3.0.0
flask-cors==4.0.0
streamlit==1.29.0
pandas==2.1.4
numpy==1.26.2
plotly==5.18.0
prophet==1.1.5
matplotlib==3.7.1
scikit-learn==1.3.2
apscheduler==3.10.4
requests==2.31.0
yfinance==0.2.33
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
reportlab==4.0.7
